refactor(fed): differentiate between setup and admin experiences
