+++ apps/guardian-ui/src/App.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useMemo, useState } from 'react';
 import { Box, VStack, Spinner, Heading, Text, Center } from '@chakra-ui/react';
 import { theme, Fonts, SharedChakraProvider } from '@fedimint/ui';
 import { GuardianApi } from './GuardianApi';
-import { GuardianProvider } from './GuardianContext';
-import { Setup } from './components/Setup';
+import { SetupContextProvider } from './SetupContext';
+import { FederationSetup } from './FederationSetup';
 import { formatApiErrorMessage } from './utils/api';
 
 export const App = React.memo(function App() {
@@ -26,36 +26,36 @@ export const App = React.memo(function App() {
     <React.StrictMode>
       <Fonts />
       <SharedChakraProvider theme={theme}>
-        <GuardianProvider api={api}>
-          <Center>
-            <Box
-              maxW='960px'
-              width='100%'
-              mt={10}
-              mb={10}
-              mr={[2, 4, 6, 10]}
-              ml={[2, 4, 6, 10]}
-              p={5}
-            >
-              {isConnected ? (
-                <Setup />
-              ) : error ? (
-                <Center>
-                  <VStack>
-                    <Heading>Something went wrong.</Heading>
-                    <Text>{error}</Text>
-                  </VStack>
-                </Center>
-              ) : (
-                <Center>
-                  <Box p={10}>
-                    <Spinner size='xl' />
-                  </Box>
-                </Center>
-              )}
-            </Box>
-          </Center>
-        </GuardianProvider>
+        <Center>
+          <Box
+            maxW='960px'
+            width='100%'
+            mt={10}
+            mb={10}
+            mr={[2, 4, 6, 10]}
+            ml={[2, 4, 6, 10]}
+            p={5}
+          >
+            {isConnected ? (
+              <SetupContextProvider api={api}>
+                <FederationSetup />
+              </SetupContextProvider>
+            ) : error ? (
+              <Center>
+                <VStack>
+                  <Heading>Something went wrong.</Heading>
+                  <Text>{error}</Text>
+                </VStack>
+              </Center>
+            ) : (
+              <Center>
+                <Box p={10}>
+                  <Spinner size='xl' />
+                </Box>
+              </Center>
+            )}
+          </Box>
+        </Center>
       </SharedChakraProvider>
     </React.StrictMode>
   );
