+++ apps/guardian-ui/src/FederationSetup.tsx
@@ -8,18 +8,18 @@ import {
   VStack,
   Spinner,
 } from '@chakra-ui/react';
-import { ReactComponent as ArrowLeftIcon } from '../assets/svgs/arrow-left.svg';
-import { Header } from './Header';
-import { useGuardianContext } from '../hooks';
-import { GuardianRole, SetupProgress, SETUP_ACTION_TYPE } from '../types';
-import { RoleSelector } from './RoleSelector';
-import { SetConfiguration } from './SetConfiguration';
-import { Login } from './Login';
-import { ConnectGuardians } from './ConnectGuardians';
-import { RunDKG } from './RunDKG';
-import { VerifyGuardians } from './VerifyGuardians';
-import { SetupComplete } from './SetupComplete';
-import { FederationDashboard } from './FederationDashboard';
+import { ReactComponent as ArrowLeftIcon } from './assets/svgs/arrow-left.svg';
+import { Header } from './components/Header';
+import { useSetupContext } from './hooks';
+import { GuardianRole, SetupProgress, SETUP_ACTION_TYPE } from './types';
+import { RoleSelector } from './components/RoleSelector';
+import { SetConfiguration } from './components/SetConfiguration';
+import { Login } from './components/Login';
+import { ConnectGuardians } from './components/ConnectGuardians';
+import { RunDKG } from './components/RunDKG';
+import { VerifyGuardians } from './components/VerifyGuardians';
+import { SetupComplete } from './components/SetupComplete';
+import { FederationAdmin } from './FederationAdmin';
 
 const PROGRESS_ORDER: SetupProgress[] = [
   SetupProgress.Start,
@@ -30,7 +30,7 @@ const PROGRESS_ORDER: SetupProgress[] = [
   SetupProgress.SetupComplete,
 ];
 
-export const Setup: React.FC = () => {
+export const FederationSetup: React.FC = () => {
   const {
     state: {
       progress,
@@ -41,7 +41,7 @@ export const Setup: React.FC = () => {
       isSetupComplete,
     },
     dispatch,
-  } = useGuardianContext();
+  } = useSetupContext();
 
   const isHost = role === GuardianRole.Host;
   const progressIdx = PROGRESS_ORDER.indexOf(progress);
@@ -71,7 +71,7 @@ export const Setup: React.FC = () => {
     subtitle = 'Please enter your password.';
     content = <Login />;
   } else if (isSetupComplete) {
-    content = <FederationDashboard />;
+    content = <FederationAdmin />;
   } else {
     switch (progress) {
       case SetupProgress.Start:
