+++ apps/guardian-ui/src/components/Login.tsx
@@ -7,12 +7,12 @@ import {
   VStack,
   FormErrorMessage,
 } from '@chakra-ui/react';
-import { useGuardianContext } from '../hooks';
+import { useSetupContext } from '../hooks';
 import { SETUP_ACTION_TYPE } from '../types';
 import { formatApiErrorMessage } from '../utils/api';
 
 export const Login: React.FC = () => {
-  const { api, dispatch } = useGuardianContext();
+  const { api, dispatch } = useSetupContext();
   const [password, setPassword] = useState('');
   const [error, setError] = useState<string>();
