+++ apps/guardian-ui/src/components/SetConfiguration.tsx
@@ -18,7 +18,7 @@ import {
   NumberInputStepper,
 } from '@chakra-ui/react';
 import { FormGroup, FormGroupHeading } from '@fedimint/ui';
-import { useGuardianContext } from '../hooks';
+import { useSetupContext } from '../hooks';
 import { BitcoinRpc, ConfigGenParams, GuardianRole, Network } from '../types';
 import { ReactComponent as FedimintLogo } from '../assets/svgs/fedimint.svg';
 import { ReactComponent as BitcoinLogo } from '../assets/svgs/bitcoin.svg';
@@ -40,7 +40,7 @@ export const SetConfiguration: React.FC<Props> = ({ next }: Props) => {
     },
     api,
     submitConfiguration,
-  } = useGuardianContext();
+  } = useSetupContext();
   const theme = useTheme();
   const isHost = role === GuardianRole.Host;
   const [myName, setMyName] = useState(stateMyName);
