+++ apps/guardian-ui/src/components/VerifyGuardians.tsx
@@ -14,7 +14,7 @@ import {
   useTheme,
 } from '@chakra-ui/react';
 import { CopyInput, FormGroup, Table } from '@fedimint/ui';
-import { useGuardianContext } from '../hooks';
+import { useSetupContext } from '../hooks';
 import { GuardianRole, Peer } from '../types';
 import { ReactComponent as ArrowRightIcon } from '../assets/svgs/arrow-right.svg';
 import { ReactComponent as CopyIcon } from '../assets/svgs/copy.svg';
@@ -34,7 +34,7 @@ export const VerifyGuardians: React.FC<Props> = ({ next }) => {
   const {
     api,
     state: { role, numPeers },
-  } = useGuardianContext();
+  } = useSetupContext();
   const theme = useTheme();
   const isHost = role === GuardianRole.Host;
   const [myHash, setMyHash] = useState('');
